Zabbix - SNMP Builder
SNMP Builder is an add-in for Zabbix. It provides new web interface
components to browse MIB trees and values. SNMP OIDs can then be
automatically converted into Zabbix items and inserted into a template.
The underlying snmpbuilder script uses calls to NetSNMP in order to
communicate with devices on the network.